Added checks for "None" directories, commands; added version to help; added CSV file output; added catches for failures in initialization; added sed tool in toolbox

+++ herptest/run_test_suite.py
@@ -10,14 +10,18 @@ import sys
 import subprocess
 import time
 import os.path
+import traceback
+import numbers
 
 from . import toolbox
 from concurrent import futures
+from edtech import sheet
+
+VERSION = '0.9.7'
 
 cfg = argparse.Namespace()
 cfg.runtime = argparse.Namespace()
 
-stdpipe = toolbox.PipeSet()
 errpipe = toolbox.PipeSet()
 
 
@@ -25,9 +29,9 @@ errpipe = toolbox.PipeSet()
 def fill_defaults():
     global cfg
 
-    if not cfg.build.prep_cmd:
+    if not hasattr(cfg.build, 'prep_cmd'):
         cfg.build.prep_cmd = [ "cmake" ]
-    if not cfg.build.compile_cmd:
+    if not hasattr(cfg.build, 'compile_cmd'):
         cfg.build.compile_cmd = [ "make", "all" ]
 
 
@@ -44,9 +48,10 @@ def parse_arguments():
     parser.add_help = True
     parser.add_argument('suite_path', nargs='?', default="./", help='path of test suite to load')
     parser.add_argument('target_path', nargs='?', default="Projects", help='path of the target projects to consider (by subdirectory / folder)')
-    parser.add_argument('-V', '--version', action='version', version='%(prog)s 0.1')
+    parser.add_argument('-V', '--version', action='version', version='%(prog)s ' + VERSION)
     parser.add_argument('-q', '--quiet', dest='quiet', action='store_true', help='execute in quiet mode')
     parser.add_argument('-d', '--debug', dest='debug', action='store_true', help='display debug information')
+    parser.add_argument('-s', '--set', dest='set', default="*", help = 'test only projects designated (e.g., *_LATE*')
     cfg.runtime = parser.parse_args(sys.argv[1:], cfg.runtime)
 
     # debug program
@@ -58,19 +63,26 @@ def parse_arguments():
 
 def build_project(source_root, build_root, prepare_cmd, compile_cmd):
     result_error = None
-
-    # If it exists, remove old project folder
-    if os.path.isdir(build_root):
-        shutil.rmtree(build_root)
-    os.makedirs(build_root)
-
-    # Build the project
     current_dir = os.getcwd()
-    os.chdir(build_root)
+
+    # If there is a not a specified build directory, fall back to the source directory instead.
+    if not build_root:
+        build_root = source_root
+
+    if build_root:
+        # If the build exists, remove old project folder, recreate it, and switch to it.
+        if os.path.isdir(build_root):
+            shutil.rmtree(build_root)
+
+        os.makedirs(build_root)
+        os.chdir(build_root)
 
     try:
-        subprocess.check_output(prepare_cmd + [source_root], stderr=subprocess.STDOUT)
-        subprocess.check_output(compile_cmd, stderr=subprocess.STDOUT)
+        if prepare_cmd:
+            subprocess.check_output(prepare_cmd + ([source_root] if source_root else []), stderr=subprocess.STDOUT)
+        if compile_cmd:
+            subprocess.check_output(compile_cmd, stderr=subprocess.STDOUT)
+
     except (subprocess.CalledProcessError, FileNotFoundError) as error:
         result_error = error
 
@@ -84,35 +96,35 @@ def run_suite_tests(framework, subject, proj_settings):
     # Run each project's tests.
     for project in proj_settings.projects:
         display_name, identifier, points = project
-        stdpipe.println("Running project " + display_name + "...")
-        score, penalty_totals = run_project_tests(identifier, framework, subject, proj_settings)
+        data_set, score, penalty_totals = run_project_tests(identifier, framework, subject, proj_settings)
 
         # If the project didn't compile, just print out a single line indicating that.
-        if score == None:
-            stdpipe.println("\nGrade:\t0 (Does not compile)\n")
+        if data_set == None:
+            data_set = [ [] ["Grade: 0 (Does not compile / run)"] ]
             score = 0
 
         else:
             # Print out info (for debugging purposes) on the score and penalty values for the project.
-            stdpipe.println("\nTest Cases:\t" + str(score * points) + " (" + str(score * 100) + "%)\n")
+            data_set += [ [], ["Test Cases: %.2f (%.2f%%)" % (score * points, score * 100) ] ]
             overall_penalty = 0
 
             for penalty_num in range(0, len(proj_settings.test_case_penalties)):
                 penalty_name, magnitude = proj_settings.test_case_penalties[penalty_num]
                 overall_penalty += penalty_totals[penalty_num]
-                stdpipe.println(penalty_name + " Penalty (overall):\t" + str(penalty_totals[penalty_num] * 100) + "%")
+                data_set += [ [ "%s Penalty (overall): %.2f%%" % (penalty_name, penalty_totals[penalty_num] * 100) ] ]
 
             for penalty_num in range(0, len(proj_settings.project_penalties)):
                 penalty_name, magnitude = proj_settings.project_penalties[penalty_num]
-                penalty_index = penalty_num + len(proj_settings.test_case_penalties)
-                overall_penalty += penalty_totals[penalty_index]
-                stdpipe.println(penalty_name + " Penalty (overall):\t" + str(penalty_totals[penalty_index] * 100) + "%")
+                penalty_ind = penalty_num + len(proj_settings.test_case_penalties)
+                overall_penalty += penalty_totals[penalty_ind]
+                data_set += [ [ "%s Penalty (overall): %.2f%%" % (penalty_name, penalty_totals[penalty_ind] * 100) ] ]
 
             # Apply the penalties and scale to the number of points
             score = (score - min(proj_settings.max_penalty, overall_penalty)) * points
 
         # Add to the results list.
-        results.append((display_name, score))
+        data_set = [ ["Project %s" % display_name] ] + data_set
+        results.append((display_name, score, data_set))
 
     return results
 
@@ -123,38 +135,68 @@ def run_project_tests(name, framework, subject, proj_settings):
     num_of_tests = proj_settings.get_number_of_tests(context)
 
     if num_of_tests == 0:
-        return None, None
+        proj_settings.shutdown_project(context)
+        return None, None, None
 
-    stdpipe.println("Number of tests: " + str(num_of_tests) + ".")
+    # Add the initial notes at the top (might adjust this later for 'pure' CSV output)
+    data_set = [ [ "Number of tests: %d." % num_of_tests ], [] ]
+
+    # Prepare the header.
+    header = [ 'Test No.', 'Score', 'Message', 'Desc.' ]
+    header.extend(["%s-Pen" % name for name, pct in proj_settings.test_case_penalties])
+    data_set.append(header)
 
     score = 0
     for test_num in range(0, num_of_tests):
-        stdpipe.print("Test case " + str(test_num) + "... ")
-        case_score = proj_settings.run_case_test(test_num, context)
-        score += case_score
-        stdpipe.print("Score: " + str(case_score * 100) + "%")
+        # Set up the row for this test and run it.
+        row = [ '%d' % test_num ]
+        case_result = proj_settings.run_case_test(test_num, context)
 
-        if case_score < 1:
-            stdpipe.print("\t[" + proj_settings.get_test_description(test_num, context) + "]")
+        # If we successfuly completed the run, this should be a number; otherwise, a message.
+        if isinstance(case_result, numbers.Number):
+            case_score = case_result
+            message = None
+        else:
+            case_score = 0
+            message = case_result
+
+        score += case_score
+        # Add score, run message, and description as applicable
+        row.append('%.2f%%' % (case_score * 100))
+        row.append(message if message else '')
+        row.append(proj_settings.get_test_description(test_num, context) if case_score < 1 else '')
 
         if case_score == 0:
-            stdpipe.println(".")
+            data_set.append(row)
             continue
 
         for penalty_num in range(0, len(proj_settings.test_case_penalties)):
-            penalty_name, magnitude = proj_settings.test_case_penalties[penalty_num]
+            pen_result = proj_settings.test_case_penalties[penalty_num]
             penalty = proj_settings.run_case_penalty(penalty_num, test_num, context)
-            penalty_totals[penalty_num] += penalty * magnitude * case_score / num_of_tests
-            stdpipe.print(";\t" + penalty_name + " Penalty (of max): " + str(penalty * 100) + "%")
 
-        stdpipe.println(".")
+            # If we successfuly completed the run, this should be a tuple; otherwise, a message.
+            if isinstance(pen_result, tuple):
+                penalty_name, magnitude = pen_result
+                message = None
+            else:
+                penalty_name = None
+                maginitude = 0
+                message = pen_result
+
+            penalty_totals[penalty_num] += penalty * magnitude * case_score / num_of_tests
+            row.append('%.2f%%' % (penalty * 100))
+
+        # Add this test data to the data set.
+        data_set.append(row)
 
     for penalty_num in range(0, len(proj_settings.project_penalties)):
         penalty_name, magnitude = proj_settings.project_penalties[penalty_num]
         penalty = proj_settings.run_project_penalty(penalty_num, context) * score * magnitude
         penalty_totals[penalty_num + len(proj_settings.test_case_penalties)] = penalty
 
-    return score / num_of_tests, penalty_totals
+    # Return the data set, score (proportion), and penalty totals
+    proj_settings.shutdown_project(context)
+    return data_set, score / num_of_tests, penalty_totals
 
 
 def make_build_paths_absolute(settings):
@@ -183,17 +225,30 @@ def prepare_and_test_submission(framework_context, submission):
     shutil.copytree(submission, cfg.build.destination, dirs_exist_ok=True)
 
     # Build the project.
-
     info("Building project(s) for " + submission + "... ")
     result_error = build_project(cfg.build.subject_src, cfg.build.subject_bin, cfg.build.prep_cmd, cfg.build.compile_cmd)
     if result_error:
-        info("error building (see logs). ")
-        errpipe.print(type(result_error).__name__ + ": " + str(result_error) + "\n")
+        info("error building (see logs)... ")
+        errpipe.print(traceback.format_exc())
 
     info("Initializing... ")
-    subject_context = cfg.project.initialize_subject(cfg.build.subject_bin)
+    subject_context = None
+    starting_dir = os.getcwd()
+
+    try:
+        subject_context = cfg.project.initialize_subject(cfg)
+    except Exception as e:
+        errpipe.print("Error initializing subject " + submission + " - " + traceback.format_exc())
+        info("error initializing (see logs)... ")
+
+    os.chdir(starting_dir)
     info("done.\n")
-    return run_suite_tests(framework_context, subject_context, cfg.project)
+
+    starting_dir = os.getcwd()
+    results = run_suite_tests(framework_context, subject_context, cfg.project)
+    cfg.project.shutdown_subject(subject_context)
+    os.chdir(starting_dir)
+    return results
 
 
 def main():
@@ -212,34 +267,34 @@ def main():
 
     cfg.project = config.project
     cfg.build = config.build
-    cfg.result_path = config.result_path
-    cfg.result_log = config.result_log
-    cfg.err_suffix = config.err_suffix
+    cfg.general = config.general
     fill_defaults()
 
     # Prepare paths
     make_build_paths_absolute(cfg.build)
-    cfg.result_path = os.path.abspath(cfg.result_path) if cfg.result_path else None
+    cfg.general.result_path = os.path.abspath(cfg.general.result_path) if cfg.general.result_path else None
 
-    if not os.path.isdir(cfg.result_path):
-        os.mkdir(cfg.result_path)
+    if not os.path.isdir(cfg.general.result_path):
+        os.mkdir(cfg.general.result_path)
 
     # Build the environment components (only need to do this once.)
     if cfg.build.framework_src and cfg.build.framework_bin:
-        info("Building framework environment... ")
-        result_error = build_project(cfg.build.framework_src, cfg.build.framework_bin, cfg.build.prep_cmd, cfg.build.compile_cmd)
-        if result_error:
-            info(type(result_error).__name__ + ": " + str(result_error) + "\n")
-            return
+        if cfg.build.prep_cmd or cfg.build.compile_cmd:
+            info("Building framework environment... ")
+            result_error = build_project(cfg.build.framework_src, cfg.build.framework_bin, cfg.build.prep_cmd, cfg.build.compile_cmd)
+            if result_error:
+                info(type(result_error).__name__ + ": " + str(result_error) + "\n")
+                return
+
         info("initializing framework... ")
-        framework_context = cfg.project.initialize_framework(cfg.build.framework_bin)
+        framework_context = cfg.project.initialize_framework(cfg)
         info("done.\n")
     else:
         framework_context = None
 
     # Prepare and run each submission.
-    for submission in glob.glob(os.path.join(cfg.runtime.target_path, "*")):
-        output_dir = os.path.join(cfg.result_path, os.path.basename(submission))
+    for submission in glob.glob(os.path.join(cfg.runtime.target_path, cfg.runtime.set)):
+        output_dir = os.path.join(cfg.general.result_path, os.path.basename(submission))
 
         if not os.path.isdir(output_dir):
             os.mkdir(output_dir)
@@ -249,26 +304,28 @@ def main():
             try:
                 suite_results = future.result()
             except Exception as e:
-                errpipe.print("Error preparing / running " + submission + " - " + type(e).__name__ + ": " + str(e))
-                toolbox.data_to_file(errpipe.read(), os.path.join(output_dir, cfg.result_log + cfg.err_suffix))
+                errpipe.print("Error preparing / running " + submission + " - " + traceback.format_exc())
+                toolbox.data_to_file(errpipe.read(), os.path.join(output_dir, cfg.general.error_log))
                 continue
 
         # Track the scores.
         grand_total = 0.0
 
-        stdpipe.println("Scores for " + submission + ":")
+        file_data = [ [ "Scores for %s..." % os.path.basename(submission) ] ]
+
         if suite_results:
-            for name, result in suite_results:
-                stdpipe.println(name + ": " + str(result))
+            for name, result, data_set in suite_results:
+                file_data.extend(data_set + [ [ "%s: %.3f" % (name, result) ] ])
                 grand_total += result
 
-        stdpipe.println("Overall score: " + str(grand_total))
+        file_data.append([ "Overall score: %.2f" % grand_total ])
         errpipe.println("")
 
-        toolbox.data_to_file(stdpipe.read(), os.path.join(output_dir, cfg.result_log))
-        toolbox.data_to_file(errpipe.read(), os.path.join(output_dir, cfg.result_log + cfg.err_suffix))
+        sheet.saveCsv(os.path.join(output_dir, cfg.general.result_file), file_data)
+        toolbox.data_to_file(errpipe.read(), os.path.join(output_dir, cfg.general.error_log))
         time.sleep(2)
 
+    cfg.project.shutdown_framework(framework_context)
     # Return to where we started at.
     os.chdir(starting_dir)
